added an ignored_first parameter to solve_heuristic
now, in addition to getting the turn order right, it can ignore certain parameters on the first move
(and only the first move).

this is kind of a hack that allows you to extract a ranked order of recommendations by artificially prohibiting the 0th (1st?) player from picking a certain player. We're not just removing the chosen player from the pool though, other players are free to pick up the artificially prohibited player (thus enabling the algorithm to react accordingly).